+++ src/lib/date.ts
@@ -1,0 +1,10 @@
+export function formatDateToYmd(date: Date): string {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}-${month}-${day}`;
+}
+
+export function getTodayYmd(): string {
+  return formatDateToYmd(new Date());
+}
